Started to implement templates (new branch).
I have a good idea of where it's going but I'm not completely sure that it will work. There are still lots of things to get done. The major one is a child-recursive clone (a clone that will go through all the child nodes and clone them as well).

--HG--
branch : fc_templates