Common selectors can be unified using the --unifyselectors.
This will simplyfy the output CSS size. This is useful when multiple parts of
the source (either directly or via an include file) have properties set for
the same set of selectors. Instead of having the selectors repeated, and showing
different properties, these are combined, keeping the output CSS very simple.